+++ libs/pandautils/src/main/java/com/instructure/pandautils/features/offline/sync/OfflineSyncHelper.kt
@@ -71,7 +71,6 @@ class OfflineSyncHelper(
 
     fun cancelRunningWorkers() {
         workManager.cancelAllWorkByTag(CourseSyncWorker.TAG)
-        workManager.cancelAllWorkByTag(FileSyncWorker.TAG)
     }
 
     private suspend fun isWorkScheduled(): Boolean {
